Small changes for getweekday macro, initial scripting for chii town
