+++ data/event_scripts.s
@@ -1156,7 +1156,7 @@ EventScript_VsSeekerChargingDone::
 	.include "data/text/save.inc"
 	.include "data/text/birch_speech.inc"
 
-	.include "data/scripts/days/sunday.inc"
+	.include "data/scripts/schedule/chii_town.inc"
 
 	.include "data/maps/SunriseVillage/scripts.inc"
