+++ data/maps/BeachboundRoute/scripts.pory
@@ -1,7 +1,6 @@
 const FLAG_HIDE_HOPPIP = FLAG_TEMP_5
 const FLAG_HIDE_SHEPHERD = FLAG_TEMP_6
 const FLAG_HIDE_KIDS = FLAG_TEMP_7
-const VAR_TIME_OF_DAY = VAR_0x8000
 const LOCALID_HOPPIP = 10
 const LOCALID_GRUNT_1 = 23
 const LOCALID_GRUNT_2 = 26
