+++ include/constants/rtc.h
@@ -1,14 +1,11 @@
+#include "vars.h"
+
 #define TIME_DEAD_NIGHT         0
 #define TIME_EARLY_MORNING      1
 #define TIME_MORNING            2
 #define TIME_LUNCHTIME          3
 #define TIME_NOONTIME           4
 #define TIME_EVENING            5
-#define TIME_NIGHTTIME          6
+#define TIME_NIGHT          6
 
-#define VAR_HOUR_OF_DAY    VAR_0x8000
-
-#define VAR_DAY            VAR_0x8000
-
-#define VAR_HOUR           VAR_0x8000
-#define VAR_MINUTE         VAR_0x8001
+#define VAR_TIME_OF_DAY         VAR_0x8000
