+++ include/fake_rtc.h
@@ -12,6 +12,5 @@ u32 FakeRtc_GetSecondsRatio(void);
 void FakeRtc_SetNewGameDay(void);
 void FakeRtc_ForwardTimeTo(u32 hour, u32 minute, u32 second);
 void PrintTimesIntoMgbaPrintf(void);
-u32 FakeRtc_GetWeekday(void);
 
 #endif // GUARD_FAKE_RTC_UTIL_H
