+++ src/rtc.c
@@ -326,21 +326,19 @@ bool8 IsBetweenHours(s32 hours, s32 begin, s32 end)
 u8 GetTimeOfDay(void)
 {
     RtcCalcLocalTime();
-    if (IsBetweenHours(hour, DEAD_NIGHT_HOUR_BEGIN, DEAD_NIGHT_HOUR_END))
+    if (IsBetweenHours(gLocalTime.hours, DEAD_NIGHT_HOUR_BEGIN, DEAD_NIGHT_HOUR_END))
         return TIME_DEAD_NIGHT;
-    else if (IsBetweenHours(hour, EARLY_MORNING_HOUR_BEGIN, EARLY_MORNING_HOUR_END))
-        return TIME_EARLY_MORNING;
-    else if (IsBetweenHours(hour, MORNING_HOUR_BEGIN, MORNING_HOUR_END))
+    else if (IsBetweenHours(gLocalTime.hours, MORNING_HOUR_BEGIN, MORNING_HOUR_END))
         return TIME_MORNING;
-    else if (IsBetweenHours(hour, LUNCHTIME_HOUR_BEGIN, LUNCHTIME_HOUR_END))
+    else if (IsBetweenHours(gLocalTime.hours, LUNCHTIME_HOUR_BEGIN, LUNCHTIME_HOUR_END))
         return TIME_LUNCHTIME;
-    else if (IsBetweenHours(hour, NOONTIME_HOUR_BEGIN, NOONTIME_HOUR_END))
+    else if (IsBetweenHours(gLocalTime.hours, NOONTIME_HOUR_BEGIN, NOONTIME_HOUR_END))
         return TIME_NOONTIME;
-    else if (IsBetweenHours(hour, EVENING_HOUR_BEGIN, EVENING_HOUR_END))
+    else if (IsBetweenHours(gLocalTime.hours, EVENING_HOUR_BEGIN, EVENING_HOUR_END))
         return TIME_EVENING;
-    else if (IsBetweenHours(hour, NIGHTTIME_HOUR_BEGIN, NIGHTTIME_HOUR_END))
-        return TIME_NIGHTTIME;
-    return NULL;
+    else if (IsBetweenHours(gLocalTime.hours, NIGHTTIME_HOUR_BEGIN, NIGHTTIME_HOUR_END))
+        return TIME_NIGHT;
+    return TIME_EARLY_MORNING;
 }
 
 void RtcInitLocalTimeOffset(s32 hour, s32 minute)
